create a file by page
