+++ apps/delivery/delivery-e2e/src/support/add-template-modal.po.ts
@@ -1,0 +1,18 @@
+/// <reference types="cypress" />
+
+import { TemplateListPage } from "./template-list.po";
+
+export class AddTemplateModal {
+  constructor() {
+    cy.contains('Add a new template');
+  }
+
+  public fillTemplateName(name: string) {
+    cy.get('input[placeholder="New template name"]').type(name);
+  }
+
+  public clickCreate() {
+    cy.get('button').contains('Create').click();
+    return new TemplateListPage();
+  }
+}
