+++ apps/delivery/delivery-e2e/src/support/delivery-teamwork.po.ts
@@ -1,0 +1,20 @@
+/// <reference types="cypress" />
+
+import { DeliveryFormPage } from "./delivery-form.po";
+
+export class DeliveryTeamWorkPage {
+  constructor() {}
+
+    public clickAddStakeholder(name: string) {
+      cy.wait(500);
+      cy.get('mat-card.mat-elevation-z0')
+      .should('contain', name)
+      .contains(name).parent().parent()
+      .find('button').click();
+    }
+
+    public clickDelivery() {
+      cy.get('a').contains('edit').click();
+      return new DeliveryFormPage;
+    }
+  }
