+++ apps/delivery/delivery-e2e/src/support/movie-teamwork.po.ts
@@ -1,0 +1,13 @@
+/// <reference types="cypress" />
+
+import { HomePage } from "./home.po";
+
+export class MovieTeamWorkPage {
+  constructor() {
+  }
+
+  public clickHome() {
+    cy.get('a').contains('home').click();
+    return new HomePage;
+  }
+}
